[MINOR] bug fixes in the GPU backend
- Each thread is assigned a cuda library handle
- JCudaKernels is also made thread safe
- Removed setting GPUContext to null
- Bug fix in initial gpu budget estimate
- Cuda Kernels use blockId.x and threadId.x only

Closes #607

+++ src/main/java/org/apache/sysml/runtime/instructions/gpu/context/GPUContextPool.java
@@ -130,7 +130,7 @@ public class GPUContextPool {
 		// initially available memory is set to the GPU with the lowest memory
 		// This is because at runtime, we wouldn't know which GPU a certain
 		// operation gets scheduled on
-		long minAvailableMemory = Integer.MAX_VALUE;
+		long minAvailableMemory = Long.MAX_VALUE;
 		for (GPUContext gCtx : pool) {
 			gCtx.initializeThread();
 			minAvailableMemory = Math.min(minAvailableMemory, gCtx.getAvailableMemory());
